test(jobs): the FleetWide gate names both River client spellings and claims only the cover it gives
Pinning ClientFromContext exactly would report a dispatcher resolving the
Safely variant as never fanning out — the spelling four call sites use and
two comments steer the next author toward — so the resolver matches as a
prefix, with a shape test holding it.

The doc comment credited jobbinding_test.go and the RLS lane with catching
a store-mediated fleet write. Neither does: one catches self-binding, the
other catches only unbound writes, and a loop that binds the GUC per tenant
and then writes satisfies RLS. That shape fails on the fan-out requirement
and nowhere else, which is what the comment now says.

Findings sort by source position so a red run reads the same twice, the
shape fixtures move to their own file, and the neighbouring wire-key gate's
counts and floors catch up with the tree they describe.
