+++ STATUS.md
@@ -311,7 +311,10 @@ lands as `workspace_id` in `river_job.args` through River's own encoder.
 means a dispatcher and nothing else: `embed_reindex` is now a dispatcher over a
 per-workspace `embed_reindex_workspace` worker, like every other fleet pass.
 `backend/jobfleetwide_test.go` holds the other half of that — a kind declaring
-`FleetWide` must actually fan out, and must issue no tenant write of its own.
+`FleetWide` must actually fan out, through one of a closed set of spellings, and
+must issue no inline SQL write in its own worker's methods. The fan-out arm is
+the load-bearing one: a worker that loops the fleet calling a store per tenant
+satisfies RLS and binds every GUC, and fails only here.
 
 **No transition was written for rows queued under the old key, deliberately.**
 They decode to a zero workspace, the binding guard refuses them before any
